third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-nourl-blocked-expected.txt

CONSOLE ERROR: line 7: Refused to load '' (MIME type 'application/x-blink-test-plugin') because it violates the following Content Security Policy Directive: 'plugin-types text/plain'.

This test passes if there is a console message saying the plugin was blocked. 
